fix: The init position of the bullets is corrected

+++ src/renderer.cpp
@@ -56,8 +56,8 @@ void Renderer::Render(Snake const snake, Snake const snake2, Shot const bullet, 
   // Render bullet1
   if (snake.shotActive == true){
     SDL_SetRenderDrawColor(sdl_renderer, 0x5F, 0x18, 0x00, 0xFF);
-    block.x = bullet.pos_x* block.w;
-    block.y = bullet.pos_y* block.h;
+    block.x = static_cast<int>(bullet.pos_x) * block.w;
+    block.y = static_cast<int>(bullet.pos_y) * block.h;
     SDL_RenderFillRect(sdl_renderer, &block);
   }
